Remove the 'we should never get here' safety check.
This check is at the end of a try/catch where both the try and the catch
will force an exit of the plugin. When performing unit tests on a plugin
we must patch the exit function so that the Python vm doesn't exit
before the tests complete. However, because of this final 'we should
never get here' check this is not the case. Instead, it issues an error
and updates the repsonse level to critical.